services/query/requirements.txt
fastapi==0.125.0
uvicorn==0.38.0
snowflake-snowpark-python==1.44.0
pandas==2.3.3
pyarrow==22.0.0
python-dotenv==1.2.1
